move MessageCache to node.py for clearer naming and less ambiguity
